Setup LALRPOP and cover basic expressions and statements

+++ compiler/src/phases/middle_end/ir.rs
@@ -167,7 +167,7 @@ pub enum ExpressionKind {
     // Primitives
     Boolean(bool),
     Identifier(Identifier),
-    Integer(i64),
+    Integer(i32),
     String(String),
     Block(Box<Block>),
     JsBlock(Type, Vec<Expression>),
